requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiohttp==3.9.1
ffmpeg-python==0.2.0
m3u8==4.0.0
sse-starlette==1.8.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
boto3>=1.28.0
